chore: move configs to separate directory
Use module build in examples.
