+++ config/rollup.js
@@ -1,0 +1,40 @@
+import commonjs from '@rollup/plugin-commonjs';
+import resolve from '@rollup/plugin-node-resolve';
+import sourcemaps from 'rollup-plugin-sourcemaps';
+import virtual from '@rollup/plugin-virtual';
+
+const virtualModules = {
+    './getXMLHttpRequest': `
+    export default function getXMLHttpRequest() {
+        return new XmlHTTPRequest();
+    }`,
+    'domain': 'export default undefined;',
+};
+
+const resolveOptions = { preferBuiltins: false };
+
+export const plugins = [
+    sourcemaps(),
+    virtual(virtualModules),
+    resolve(resolveOptions),
+    commonjs(),
+];
+
+export const umdOutput = {
+    format: 'umd',
+    name: 'Mapillary',
+    sourcemap: true,
+}
+
+export const srcInput = 'build/esm/src/Mapillary.js';
+
+export const esm = {
+    input: srcInput,
+    output: [
+        {
+            file: 'dist/mapillary.module.js',
+            format: 'es',
+            sourcemap: true,
+        }],
+    plugins,
+};
