+++ config/rollup.serve.config.js
@@ -1,0 +1,3 @@
+import { esm } from './rollup.js';
+
+export default esm;
