+++ package.json
@@ -78,22 +78,22 @@
   ],
   "scripts": {
     "build": "yarn build-styles && yarn build-src",
-    "build-docs": "typedoc --options typedoc.json",
+    "build-docs": "typedoc --options ./config/typedoc.json",
     "build-src": "yarn compile && yarn bundle",
-    "build-styles": "cat styles/*.css | postcss -o dist/mapillary.css",
+    "build-styles": "cat styles/*.css | postcss --config ./config -o dist/mapillary.css",
     "bundle": "rollup -c",
-    "bundle-watch": "rollup -w --no-watch.clearScreen -c rollup.umd.config.js",
+    "bundle-watch": "rollup -w --no-watch.clearScreen -c ./config/rollup.serve.config.js",
     "clean": "yarn clean-build && yarn clean-dist && yarn clean-docs",
     "clean-build": "rm -rf build",
     "clean-dist": "rm -rf dist && mkdir dist",
     "clean-docs": "rm -rf docs/build",
     "compile": "tsc -p tsconfig.json",
     "compile-incremental": "tsc --incremental -p tsconfig.json",
-    "compile-test": "tsc -p tsconfig.test.json",
+    "compile-test": "tsc -p ./config/tsconfig.test.json",
     "compile-watch": "tsc -w --incremental --preserveWatchOutput -p tsconfig.json",
     "lint": "yarn lint-spec && yarn lint-src",
-    "lint-spec": "tslint -c tslint.json -p tsconfig.json \"spec/**/*.ts\"",
-    "lint-src": "tslint -c tslint.json -p tsconfig.json \"src/**/*.ts\"",
+    "lint-spec": "tslint -c ./config/tslint.json -p tsconfig.json \"spec/**/*.ts\"",
+    "lint-src": "tslint -c ./config/tslint.json -p tsconfig.json \"src/**/*.ts\"",
     "prepare": "yarn clean && yarn lint && yarn build",
     "serve": "node server.js",
     "serve-docs": "python3 -m http.server",
