+++ src/viewer/PlayService.ts
@@ -14,11 +14,9 @@ import {
     catchError,
     distinctUntilChanged,
     filter,
-    finalize,
     first,
     map,
     mergeMap,
-    publish,
     publishReplay,
     refCount,
     retry,
@@ -29,7 +27,6 @@ import {
     withLatestFrom,
 } from "rxjs/operators";
 
-import { LatLon } from "../api/interfaces/LatLon";
 import { GraphCalculator } from "../graph/GraphCalculator";
 import { GraphMode } from "../graph/GraphMode";
 import { GraphService } from "../graph/GraphService";
